refactor(anthropic): inline sampling overrides
